Implemented AsyncStreamReader interface
